1.21.130 update Fix steve glitch. Fix ghost player spawn to all bug from game mode changes. Fix rods having left over fishing hooks. Make block manager use a more secure vector key to avoid bugs with rounding and negative coordinates. Make block manager and position helper use block coordinates to avoid implicit conversions and unintended consequences of float rounding. Cleaned up a lot of forms. Boom box logic now uses underwater tnt for knocker boxes. No more random UUIDs in player list entry packets due to the client requiring the real UUIDs to receive skin packets properly.

+++ src/core/systems/scene/Scene.php
@@ -416,7 +416,7 @@ abstract class Scene
     }
   }
 
-  public function sceneSound(string $soundName, int $volume = 2, int $pitch = 1): void
+  public function sceneSound(string $soundName, float $volume = 2, float $pitch = 1): void
   {
     foreach ($this->players as $player) {
       if ($player->isOnline()) ServerSounds::playSoundToPlayer($player, $soundName, $volume, $pitch);
